+++ fmpz_mpoly/test/t-resultant_discriminant.c
@@ -14,6 +14,7 @@
 #include <gmp.h>
 #include "flint.h"
 #include "fmpz.h"
+#include "fmpz_poly.h"
 #include "fmpz_mpoly.h"
 #include "ulong_extras.h"
 #include "profiler.h"
@@ -25,52 +26,76 @@ main(void)
 
     FLINT_TEST_INIT(state);
 
-    flint_printf("resultant_discriminant....\n");
+    flint_printf("resultant_discriminant....");
     fflush(stdout);
 
 
+    /* Check univariate resultant */
+    for (i = 0; i < 50 * flint_test_multiplier(); i++)
     {
-
-
-
         fmpz_mpoly_ctx_t ctx;
-        fmpz_mpoly_t a, b, c;
-        fmpz_mpoly_univar_t ax, bx, cx;
-        timeit_t time;
-        const char* vars[] = {"x","y","z","a","b","c","d","e","f","g","h"};
+        fmpz_mpoly_t a, b, r;
+        fmpz_poly_t au, bu;
+        fmpz_t ru;
+        ordering_t ord;
+        slong nvars, len1, len2, exp_bound1, exp_bound2;
+        slong coeff_bits;
+        slong ashift, bshift;
 
-        fmpz_mpoly_ctx_init(ctx, 2, ORD_DEGREVLEX);
+        ord = mpoly_ordering_randtest(state);
+        nvars = 1;
+
+        fmpz_mpoly_ctx_init(ctx, nvars, ord);
+
         fmpz_mpoly_init(a, ctx);
         fmpz_mpoly_init(b, ctx);
-        fmpz_mpoly_init(c, ctx);
+        fmpz_mpoly_init(r, ctx);
+        fmpz_poly_init(au);
+        fmpz_poly_init(bu);
+        fmpz_init(ru);
 
-        fmpz_mpoly_set_str_pretty(a, "-x^4*y-2*x^4-y^3-x^2", vars, ctx);
-        fmpz_mpoly_set_str_pretty(b, "2*y^2", vars, ctx);
+        len1 = n_randint(state, 100);
+        len2 = n_randint(state, 100);
+        exp_bound1 = n_randint(state, 50) + 1;
+        exp_bound2 = n_randint(state, 50) + 1;
+        coeff_bits = n_randint(state, 4);
+        fmpz_mpoly_randtest(a, state, len1, exp_bound1, coeff_bits, ctx);
+        fmpz_mpoly_randtest(b, state, len2, exp_bound2, coeff_bits, ctx);
 
-        printf("A: "); fmpz_mpoly_print_pretty(a, vars, ctx); printf("\n");
-        printf("B: "); fmpz_mpoly_print_pretty(b, vars, ctx); printf("\n");
-        fmpz_mpoly_resultant(c, a, b, 1, ctx);
-        printf("C: "); fmpz_mpoly_print_pretty(c, vars, ctx); printf("\n");
+        fmpz_mpoly_to_fmpz_poly(au, &ashift, a, 0, ctx);
+        fmpz_mpoly_to_fmpz_poly(bu, &bshift, b, 0, ctx);
+        fmpz_poly_shift_left(au, au, ashift);
+        fmpz_poly_shift_left(bu, bu, bshift);
+
+        fmpz_poly_resultant(ru, au, bu);
+        fmpz_mpoly_resultant(r, a, b, 0, ctx);
+        if (!fmpz_mpoly_equal_fmpz(r, ru, ctx))
+        {
+            printf("FAIL\n");
+            flint_printf(" Check univariate resultant \ni: %wd\n",i);
+            flint_abort();
+        }
 
         fmpz_mpoly_clear(a, ctx);
         fmpz_mpoly_clear(b, ctx);
-        fmpz_mpoly_clear(c, ctx);
+        fmpz_mpoly_clear(r, ctx);
+        fmpz_poly_clear(au);       
+        fmpz_poly_clear(bu);
+        fmpz_clear(ru);
     }
 
 
-
     /* Check res(a*b,c) = res(a,c)*res(b,c) */
-    for (i = 0; i < 0*30 * flint_test_multiplier(); i++)
+    for (i = 0; i < 30 * flint_test_multiplier(); i++)
     {
         fmpz_mpoly_ctx_t ctx;
         fmpz_mpoly_t a, b, c, ab, ra, rb, rab, p;
-        fmpz_mpoly_univar_t fx;
         ordering_t ord;
-        slong nvars, len1, len2, len3, exp_bound, exp_bound1, exp_bound2, exp_bound3;
-        slong coeff_bits, exp_bits1, exp_bits2, exp_bits3;
+        slong nvars, len1, len2, len3, exp_bound1, exp_bound2, exp_bound3;
+        slong coeff_bits;
 
         ord = mpoly_ordering_randtest(state);
-        nvars = n_randint(state, 2) + 1;
+        nvars = n_randint(state, 3) + 1;
 
         fmpz_mpoly_ctx_init(ctx, nvars, ord);
 
@@ -83,12 +108,12 @@ main(void)
         fmpz_mpoly_init(rab, ctx);
         fmpz_mpoly_init(p, ctx);
 
-        len1 = n_randint(state, 20);
-        len2 = n_randint(state, 20);
-        len3 = n_randint(state, 20);
-        exp_bound1 = n_randint(state, 8) + 1;
-        exp_bound2 = n_randint(state, 8) + 1;
-        exp_bound3 = n_randint(state, 8) + 1;
+        len1 = n_randint(state, 15);
+        len2 = n_randint(state, 15);
+        len3 = n_randint(state, 15);
+        exp_bound1 = n_randint(state, 5) + 1;
+        exp_bound2 = n_randint(state, 5) + 1;
+        exp_bound3 = n_randint(state, 5) + 1;
         coeff_bits = n_randint(state, 3);
         fmpz_mpoly_randtest(a, state, len1, exp_bound1, coeff_bits, ctx);
         fmpz_mpoly_randtest(b, state, len2, exp_bound2, coeff_bits, ctx);
@@ -105,7 +130,8 @@ main(void)
             if (!fmpz_mpoly_equal(p,rab,ctx))
             {
                 printf("FAIL\n");
-                flint_printf(" Check res(a*b,c) = res(a,c)*res(b,c) \ni: %wd  j: %wd\n",i,j);
+                flint_printf(" Check res(a*b,c) = res(a,c)*res(b,c) \n"
+                                                       "i: %wd  j: %wd\n",i,j);
                 flint_abort();
             }
         }
@@ -122,16 +148,16 @@ main(void)
 
 
     /* Check disc(a*b) = disc(a)*disc(b)*res(a,b)^2 */
-    for (i = 0; i < 0*30 * flint_test_multiplier(); i++)
+    for (i = 0; i < 30 * flint_test_multiplier(); i++)
     {
         fmpz_mpoly_ctx_t ctx;
         fmpz_mpoly_t a, b, ab, r, da, db, dab, p;
         ordering_t ord;
         slong nvars, len1, len2, exp_bound1, exp_bound2;
-        slong coeff_bits, exp_bits1, exp_bits2;
+        slong coeff_bits;
 
         ord = mpoly_ordering_randtest(state);
-        nvars = n_randint(state, 2) + 1;
+        nvars = n_randint(state, 3) + 1;
 
         fmpz_mpoly_ctx_init(ctx, nvars, ord);
 
@@ -144,10 +170,10 @@ main(void)
         fmpz_mpoly_init(r, ctx);
         fmpz_mpoly_init(p, ctx);
 
-        len1 = n_randint(state, 20);
-        len2 = n_randint(state, 20);
-        exp_bound1 = n_randint(state, 8) + 1;
-        exp_bound2 = n_randint(state, 8) + 1;
+        len1 = n_randint(state, 15);
+        len2 = n_randint(state, 15);
+        exp_bound1 = n_randint(state, 5) + 1;
+        exp_bound2 = n_randint(state, 5) + 1;
         coeff_bits = n_randint(state, 3);
         fmpz_mpoly_randtest(a, state, len1, exp_bound1, coeff_bits, ctx);
         fmpz_mpoly_randtest(b, state, len2, exp_bound2, coeff_bits, ctx);
@@ -171,7 +197,8 @@ main(void)
             if (!fmpz_mpoly_equal(dab, p, ctx))
             {
                 printf("FAIL\n");
-                flint_printf(" Check disc(a*b) = disc(a)*disc(b)*res(a,b)^2 \ni: %wd  j: %wd\n",i,j);
+                flint_printf(" Check disc(a*b) = disc(a)*disc(b)*res(a,b)^2 \n"
+                                                       "i: %wd  j: %wd\n",i,j);
                 flint_abort();
             }
         }
