+++ mpoly/gcd_prs.c
@@ -1,0 +1,492 @@
+/*
+    Copyright (C) 2017 Daniel Schultz
+
+    This file is part of FLINT.
+
+    FLINT is free software: you can redistribute it and/or modify it under
+    the terms of the GNU Lesser General Public License (LGPL) as published
+    by the Free Software Foundation; either version 2.1 of the License, or
+    (at your option) any later version.  See <http://www.gnu.org/licenses/>.
+*/
+
+#include <gmp.h>
+#include <stdlib.h>
+#include "flint.h"
+#include "fmpz.h"
+#include "fmpz_poly.h"
+#include "fmpz_mpoly.h"
+#include "assert.h"
+
+
+void fmpz_mpoly_gcd_monomial(fmpz_mpoly_t poly1, const fmpz_mpoly_t polyA, const fmpz_mpoly_t polyB, const fmpz_mpoly_ctx_t ctx)
+{
+    int deg, rev;
+    slong i, N, bits;
+    ulong * texpA, * texpB, * exps;
+    ulong mask;
+    fmpz_t igcd;
+    TMP_INIT;
+
+    assert(polyA->length == 1);
+    assert(polyB->length == 1);
+
+    TMP_START;
+    texpA = (ulong *) TMP_ALLOC(ctx->n*sizeof(ulong));
+    texpB = (ulong *) TMP_ALLOC(ctx->n*sizeof(ulong));
+    exps = (ulong *) TMP_ALLOC(ctx->n*sizeof(ulong));
+
+    fmpz_init(igcd);
+    fmpz_gcd(igcd, polyA->coeffs + 0, polyB->coeffs + 0);
+
+    bits = FLINT_MAX(polyA->bits, polyB->bits);
+    N = words_per_exp(ctx->n, bits);
+    degrev_from_ord(deg, rev, ctx->ord);
+    mask = 0;
+    for (i = 0; i < FLINT_BITS/bits; i++)
+        mask = (mask << bits) + (UWORD(1) << (bits - 1));
+
+    mpoly_unpack_monomials(texpA, bits, polyA->exps, polyA->bits, 1, ctx->n);
+    mpoly_unpack_monomials(texpB, bits, polyB->exps, polyB->bits, 1, ctx->n);
+    mpoly_monomial_min(texpA, texpA, texpB, bits, N, mask);
+    mpoly_get_monomial(exps, texpA, bits, ctx->n, deg, rev);
+    
+    fmpz_mpoly_fit_length(poly1, 1, ctx);
+    fmpz_mpoly_fit_bits(poly1, bits, ctx);
+    poly1->bits = bits;
+    mpoly_set_monomial(poly1->exps + N*0, exps, bits, ctx->n, deg, rev);
+    fmpz_set(poly1->coeffs + 0, igcd);
+    _fmpz_mpoly_set_length(poly1, 1, ctx);
+
+    fmpz_clear(igcd);
+
+    TMP_END;
+}
+
+
+
+void _fmpz_mpoly_gcd_prs(fmpz_mpoly_t poly1, const fmpz_mpoly_t polyA, const fmpz_mpoly_t polyB, const fmpz_mpoly_ctx_t ctx)
+{
+
+    slong shift, off, bits, fpw, N;
+    slong i, m, c, d, v, k, var, nvars;
+    int deg, rev;
+    ulong mask;
+    slong * a_degs, * b_degs, * a_leads, * b_leads;
+
+    fmpz_mpoly_t A,B;
+
+    fmpz_mpoly_t ac, bc, gc, gabc, g;
+    fmpz_mpoly_univar_t ax, bx, gx;
+/*
+    const char* vars[] = {"x","y","z","t","a","b","c","d","e","f","g","h"};
+*/
+    TMP_INIT;
+
+    TMP_START;
+
+/*
+printf("**********\ncalled _fmpz_mpoly_gcd_prs(");
+fmpz_mpoly_print_pretty(polyA,vars,ctx);
+flint_printf(", ");
+fmpz_mpoly_print_pretty(polyB,vars,ctx);
+flint_printf(")\n");
+*/
+
+
+    fmpz_mpoly_init(A, ctx);
+    fmpz_mpoly_init(B, ctx);
+    fmpz_mpoly_set(A, polyA, ctx);
+    fmpz_mpoly_set(B, polyB, ctx);
+
+    fmpz_mpoly_init(ac, ctx);
+    fmpz_mpoly_init(bc, ctx);
+    fmpz_mpoly_init(gc, ctx);
+    fmpz_mpoly_init(gabc, ctx);
+    fmpz_mpoly_init(g, ctx);
+    fmpz_mpoly_univar_init(ax, ctx);
+    fmpz_mpoly_univar_init(bx, ctx);
+    fmpz_mpoly_univar_init(gx, ctx);
+
+
+
+
+    degrev_from_ord(deg, rev, ctx->ord);    
+    nvars = ctx->n - deg;
+
+
+    if (polyA->length == 0)
+    {
+        fmpz_mpoly_set(poly1, polyB, ctx);
+        goto done;
+    } else if (polyB->length == 0)
+    {
+        fmpz_mpoly_set(poly1, polyA, ctx);
+        goto done;
+    }
+/*
+printf("neither is zero\n");
+*/
+    if (polyA->length == 1)
+    {
+        fmpz_mpoly_term_content(bc, polyB, ctx);
+        fmpz_mpoly_gcd_monomial(poly1, polyA, bc, ctx);
+        goto done;
+    } else if (polyB->length == 1)
+    {
+        fmpz_mpoly_term_content(ac, polyA, ctx);
+        fmpz_mpoly_gcd_monomial(poly1, polyB, ac, ctx);
+        goto done;
+    }
+
+/*
+printf("neither is monomial\n");
+*/
+
+    a_degs = (slong *)TMP_ALLOC(nvars*sizeof(slong));
+    b_degs = (slong *)TMP_ALLOC(nvars*sizeof(slong));
+    a_leads = (slong *)TMP_ALLOC(nvars*sizeof(slong));
+    b_leads = (slong *)TMP_ALLOC(nvars*sizeof(slong));
+    fmpz_mpoly_degrees(a_degs, polyA, ctx);
+    fmpz_mpoly_degrees(b_degs, polyB, ctx);
+
+    d = 0;
+    for (v = 0; v < nvars; v++)
+    {
+        d |= a_degs[v] | b_degs[v];
+    }
+    if (d == 0)
+    {
+        fmpz_t gg;
+        assert(polyA->length == 1);
+        assert(polyB->length == 1);
+        fmpz_init(gg);
+        fmpz_gcd(gg, polyA->coeffs + 0, polyB->coeffs + 0);
+        fmpz_mpoly_set_fmpz(poly1, gg, ctx);
+        fmpz_clear(gg);
+        goto done;
+    }
+
+/*
+printf("at least one is non constant\n");
+*/
+
+    for (v = 0; v < nvars; v++)
+    {
+        if (a_degs[v] == 0 && b_degs[v] != 0)
+        {
+            fmpz_mpoly_to_univar(bx, polyB, v, ctx);
+            fmpz_mpoly_set(poly1, polyA, ctx);
+            for (i = 0; i < bx->length; i++)
+                _fmpz_mpoly_gcd_prs(poly1, poly1, bx->coeffs + i, ctx);
+            goto done;
+        }
+        if (b_degs[v] == 0 && a_degs[v] != 0)
+        {
+            fmpz_mpoly_to_univar(ax, polyA, v, ctx);
+            fmpz_mpoly_set(poly1, polyB, ctx);
+            for (i = 0; i < ax->length; i++)
+                _fmpz_mpoly_gcd_prs(poly1, poly1, ax->coeffs + i, ctx);
+            goto done;
+        }
+    }
+/*
+printf("matching variables\n");
+*/
+    for (v = 0; v < nvars; v++)
+    {
+        bits = polyA->bits;
+        fpw = FLINT_BITS/bits;
+        mask = (-UWORD(1)) >> (FLINT_BITS - bits);
+        N = words_per_exp(ctx->n, bits);
+        degrev_from_ord(deg, rev, ctx->ord);
+        mpoly_off_shift(&off, &shift, v, deg, rev, fpw, ctx->n, bits);
+        if (a_degs[v] != 0)
+        {
+            a_leads[v] = 0;
+            for (i = 0; i < polyA->length; i++)
+                a_leads[v] += ((polyA->exps[N*i + off] >> shift) & mask) == a_degs[v];
+        }
+
+        bits = polyB->bits;
+        fpw = FLINT_BITS/bits;
+        mask = (-UWORD(1)) >> (FLINT_BITS - bits);
+        N = words_per_exp(ctx->n, bits);
+        degrev_from_ord(deg, rev, ctx->ord);
+        mpoly_off_shift(&off, &shift, v, deg, rev, fpw, ctx->n, bits);
+        if (b_degs[v] != 0)
+        {
+            b_leads[v] = 0;
+            for (i = 0; i < polyB->length; i++)
+                b_leads[v] += ((polyB->exps[N*i + off] >> shift) & mask) == b_degs[v];
+        }
+    }
+
+/*
+for (v = 0; v < nvars; v++)
+{
+    flint_printf("%wd:  a_deg %wd  a_lead %wd   b_deg %wd  b_lead %wd\n", v, a_degs[v], a_leads[v], b_degs[v], b_leads[v]);
+}
+*/
+
+    k = 0;
+    m = 1000000;
+    for (v = 0; v < nvars; v++)
+    {
+        if (a_degs[v] != 0)
+        {
+            if (a_degs[v] >= b_degs[v])
+                c = FLINT_MAX(FLINT_BIT_COUNT(b_leads[v]-1)*a_degs[v], 1)*b_degs[v];
+            else
+                c = FLINT_MAX(FLINT_BIT_COUNT(a_leads[v]-1)*b_degs[v], 1)*a_degs[v];
+
+            if (c < m)
+            {
+                m = c;
+                k = v;
+            }
+        }
+    }
+    var = k;
+
+/*
+flint_printf("choosing variable %wd\n", var);
+*/
+
+    d = 0;
+    for (v = 0; v < nvars; v++)
+    {
+        if (v != k)
+            d |= a_degs[v] | b_degs[v];
+    }
+    if (d == 0)
+    {
+/*
+flint_printf("using univar in ");
+flint_printf("_fmpz_mpoly_gcd_prs(");
+fmpz_mpoly_print_pretty(polyA,vars,ctx);
+flint_printf(", ");
+fmpz_mpoly_print_pretty(polyB,vars,ctx);
+flint_printf(")\n");
+*/
+
+        slong e1, e2, e3;
+        fmpz_poly_t u1, u2, u3;
+        fmpz_poly_init(u1);
+        fmpz_poly_init(u2);
+        fmpz_poly_init(u3);
+        fmpz_mpoly_to_fmpz_poly(u2, &e2, polyA, k, ctx);
+        fmpz_mpoly_to_fmpz_poly(u3, &e3, polyB, k, ctx);
+/*
+printf("**************\n");
+printf("u2: "); fmpz_poly_print_pretty(u2,"X"); printf("\n");
+printf("u3: "); fmpz_poly_print_pretty(u3,"X"); printf("\n");
+*/
+        fmpz_poly_gcd(u1, u2, u3);
+/*
+printf("u1: "); fmpz_poly_print_pretty(u1,"X"); printf("\n");
+*/
+        e1 = FLINT_MIN(e2, e3);
+        fmpz_mpoly_from_fmpz_poly(poly1, u1, e1, k, ctx);
+
+
+
+
+        fmpz_poly_clear(u1);
+        fmpz_poly_clear(u2);
+        fmpz_poly_clear(u3);
+        goto done;
+    }
+
+/*
+flint_printf("using pgcd\n");
+*/
+
+    fmpz_mpoly_to_univar(ax, polyA, var, ctx);
+    fmpz_mpoly_to_univar(bx, polyB, var, ctx);
+    gx->var = var;
+
+    fmpz_mpoly_set(ac, ax->coeffs + 0, ctx);
+    for (i = 1; i < ax->length; i++)
+    {
+        _fmpz_mpoly_gcd_prs(ac, ac, ax->coeffs + i, ctx);
+    }
+    fmpz_mpoly_set(bc, bx->coeffs + 0, ctx);
+    for (i = 1; i < bx->length; i++)
+    {
+        _fmpz_mpoly_gcd_prs(bc, bc, bx->coeffs + i, ctx);
+    }
+
+    for (i = 0; i < ax->length; i++)
+    {
+        fmpz_mpoly_divides_monagan_pearce(ax->coeffs + i, ax->coeffs + i, ac, ctx);
+    }
+    for (i = 0; i < bx->length; i++)
+    {
+        fmpz_mpoly_divides_monagan_pearce(bx->coeffs + i, bx->coeffs + i, bc, ctx);
+    }
+
+    if (ax->exps[0] == 0 || bx->exps[0] == 0)
+    {
+        fmpz_mpoly_univar_fit_length(gx, 1, ctx);
+        fmpz_mpoly_set_ui(gx->coeffs + 0, WORD(1), ctx);
+        gx->exps[0] = 0;
+        gx->length = 1;
+
+    } else {
+        if (ax->exps[0] >= bx->exps[0])
+        {
+            _fmpz_mpoly_univar_pgcd(gx, ax, bx, ctx);
+        } else {
+            _fmpz_mpoly_univar_pgcd(gx, bx, ax, ctx);
+        }
+        assert(gx->length > 0);
+    }
+
+/*
+flint_printf("gx(%wd):\n",gx->length);
+fmpz_mpoly_univar_print(gx, vars, ctx);
+printf("\n");
+*/
+    assert(gx->length > 0);
+    if ((gx->coeffs + 0)->length != 1 && (gx->coeffs + gx->length - 1)->length != 1)
+    {
+/*
+flint_printf("ax: ");
+fmpz_mpoly_univar_print(ax, vars, ctx);
+printf("\n");
+flint_printf("bx: ");
+fmpz_mpoly_univar_print(bx, vars, ctx);
+printf("\n");
+*/
+
+        if ((ax->coeffs + 0)->length == 1 || (bx->coeffs + 0)->length == 1)
+        {
+            fmpz_mpoly_term_content(gc, gx->coeffs + 0, ctx);
+            fmpz_mpoly_divides_monagan_pearce(gabc, gx->coeffs + 0, gc, ctx);
+            for (i = 0; i < gx->length; i++)
+            {
+                if (!fmpz_mpoly_divides_monagan_pearce(gx->coeffs + i, gx->coeffs + i, gabc, ctx))
+                    assert(0 && "not lead div");
+            }
+        } else if ((ax->coeffs + ax->length-1)->length == 1 || (bx->coeffs + bx->length-1)->length == 1)
+        {
+            fmpz_mpoly_term_content(gc, gx->coeffs + gx->length-1, ctx);
+            fmpz_mpoly_divides_monagan_pearce(gabc, gx->coeffs + gx->length-1, gc, ctx);
+            for (i = 0; i < gx->length; i++)
+            {
+                if (!fmpz_mpoly_divides_monagan_pearce(gx->coeffs + i, gx->coeffs + i, gabc, ctx))
+                    assert(0 && "not trail div");
+            }
+        } else
+        {
+            _fmpz_mpoly_gcd_prs(gc, ax->coeffs + 0, bx->coeffs + 0, ctx);
+            if (gc->length == 1)
+            {
+                fmpz_mpoly_term_content(gc, gx->coeffs + 0, ctx);
+                fmpz_mpoly_divides_monagan_pearce(gabc, gx->coeffs + 0, gc, ctx);
+                for (i = 0; i < gx->length; i++)
+                {
+                    if (!fmpz_mpoly_divides_monagan_pearce(gx->coeffs + i, gx->coeffs + i, gabc, ctx))
+                        assert(0 && "not lead div");
+                }                
+            }
+        }
+    }
+
+
+    fmpz_mpoly_term_content(gc, gx->coeffs + 0, ctx);
+    for (i = 1; i < gx->length; i++)
+    {
+        fmpz_mpoly_term_content(gabc, gx->coeffs + i, ctx);
+        fmpz_mpoly_gcd_monomial(gc, gc, gabc, ctx);
+    }
+    for (i = 0; i < gx->length; i++)
+    {
+        fmpz_mpoly_divides_monagan_pearce(gx->coeffs + i, gx->coeffs + i, gc, ctx);
+    }
+
+    fmpz_mpoly_set(gc, gx->coeffs + 0, ctx);
+    for (i = 1; i < gx->length; i++)
+    {
+        _fmpz_mpoly_gcd_prs(gc, gc, gx->coeffs + i, ctx);
+    }
+
+    for (i = 0; i < gx->length; i++)
+    {
+        fmpz_mpoly_divides_monagan_pearce(gx->coeffs + i, gx->coeffs + i, gc, ctx);
+    }
+
+    fmpz_mpoly_from_univar(g, gx, ctx);
+
+
+    _fmpz_mpoly_gcd_prs(gabc, ac, bc, ctx);
+
+    fmpz_mpoly_mul_johnson(poly1, g, gabc, ctx);
+
+done:
+
+
+    fmpz_mpoly_test(poly1, ctx);
+    fmpz_mpoly_test(A, ctx);
+    fmpz_mpoly_test(B, ctx);
+
+/*
+    if (poly1->length > 0)
+    {
+        if (!fmpz_mpoly_divides_monagan_pearce(ac, A, poly1, ctx)
+            || !fmpz_mpoly_divides_monagan_pearce(ac, B, poly1, ctx))
+        {
+
+printf("**********\nproblem ");
+fmpz_mpoly_print_pretty(poly1,vars,ctx);
+flint_printf(" != ");
+flint_printf("_fmpz_mpoly_gcd_prs(");
+fmpz_mpoly_print_pretty(A,vars,ctx);
+flint_printf(", ");
+fmpz_mpoly_print_pretty(B,vars,ctx);
+flint_printf(")\n");
+assert(0);
+
+        }
+
+    }
+*/
+
+/*
+printf("**********\nreturning ");
+fmpz_mpoly_print_pretty(poly1,vars,ctx);
+flint_printf(" = ");
+flint_printf("_fmpz_mpoly_gcd_prs(");
+fmpz_mpoly_print_pretty(A,vars,ctx);
+flint_printf(", ");
+fmpz_mpoly_print_pretty(B,vars,ctx);
+flint_printf(")\n");
+*/
+
+    TMP_END;
+/*
+    fmpz_mpoly_clear(A, ctx);
+    fmpz_mpoly_clear(B, ctx);
+*/
+
+    fmpz_mpoly_clear(ac, ctx);
+    fmpz_mpoly_clear(bc, ctx);
+    fmpz_mpoly_clear(gc, ctx);
+    fmpz_mpoly_clear(gabc, ctx);
+    fmpz_mpoly_clear(g, ctx);
+
+    fmpz_mpoly_univar_clear(ax, ctx);
+    fmpz_mpoly_univar_clear(bx, ctx);
+    fmpz_mpoly_univar_clear(gx, ctx);
+
+}
+
+
+void fmpz_mpoly_gcd_prs(fmpz_mpoly_t poly1, fmpz_mpoly_t poly2, fmpz_mpoly_t poly3, const fmpz_mpoly_ctx_t ctx)
+{
+    _fmpz_mpoly_gcd_prs(poly1, poly2, poly3, ctx);
+    if ((poly1->length > 0) && (fmpz_cmp_ui(poly1->coeffs + 0, WORD(0)) < 0))
+        fmpz_mpoly_neg(poly1, poly1, ctx);
+}
+
+
